Prepare the editor split: dead code, style imports, import edits
Delete the unused preview/debug.rs Graphviz dump.
Give the visual editor UI its own copies of the two styling.slint
symbols it used, so ui/visual-editor no longer imports preview styling.
Move create_import_edit and its helpers from language/completion.rs to
common/import_edit.rs, since the preview needs them and language/ stays
in the LSP when the editor-preview crate is split out.
