+++ tools/lsp/common.rs
@@ -21,6 +21,7 @@ pub use document_cache::DocumentCache;
 pub use i_slint_compiler::diagnostics::ByteFormat;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod host_language_search;
+pub mod import_edit;
 mod lsp_to_previews;
 pub mod rename_component;
 pub mod rename_element_id;
