+++ tools/lsp/ui/visual-editor/components/common.slint
@@ -37,6 +37,12 @@ export global StudioTheme {
     out property <color> object-subtle: #f6f8fa;
     out property <color> object-border: #d0d7de;
     out property <color> object-muted: #6e7781;
+    out property <brush> drop-mark-background: #ffffff;
+    out property <brush> drop-mark-foreground: #ffffff;
+}
+
+export global EditorIcons {
+    out property <image> select-chevron: @image-url("../../assets/octicons/select-chevron.svg");
 }
 
 export global EditorMetrics {
