+++ tools/lsp/ui/visual-editor/components/editor-canvas.slint
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-Royalty-free-2.0 OR LicenseRef-Slint-Software-3.0
 
 import { Api, CanvasCorner, DropMark, PrototypeComponentKind, SelectionRectangle } from "../../api.slint";
-import { EditorPalette } from "../../components/styling.slint";
 import {
     CanvasOrientation, CanvasSelectionKind, EditorMetrics, ImagePlaceholder, StudioTheme,
 } from "common.slint";
@@ -194,8 +193,8 @@ export component EditorCanvas inherits Rectangle {
                 y: mark-y;
                 width: max((drop-mark.x2 - drop-mark.x1).abs(), 2px);
                 height: max((drop-mark.y2 - drop-mark.y1).abs(), 2px);
-                border-color: EditorPalette.drop-mark-foreground;
-                background: EditorPalette.drop-mark-background;
+                border-color: StudioTheme.drop-mark-foreground;
+                background: StudioTheme.drop-mark-background;
             }
 
             // Click to select the real previewed element.
